requirements.txt
# Hadith Narrator Graph - Python Dependencies

# Neo4j Python Driver
neo4j>=5.0.0,<6.0.0

# MongoDB Atlas (mongo_migration/)
pymongo[srv]>=4.6

# Environment variable management
python-dotenv>=1.0.0

# FastAPI backend
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Table formatting for CLI output
tabulate>=0.9.0

# Optional: For LLM-based extraction (app.py)
# langextract>=0.1.0
# openai>=1.0.0
